Don't set default marker offset to [0,0]
Maplibre sets its own offset by default so that the "center" point of
the marker is the bottom of the marker pin. Setting the offset to [0,0]
was ovwrriding this, causing the marker to be centered in the middle of
the icon which makes it slightly inaccurate and appear to drift as the map zooms.

Fixes #237